Changes: Rebased with master and Modified Test Cases
