--- apps/acqdat_api/lib/acqdat_api/sensor_type.ex
@@ -1,51 +1,0 @@
-defmodule AcqdatApi.SensorType do
-  alias AcqdatCore.Model.SensorType, as: SensorTypeModel
-  import AcqdatApiWeb.Helpers
-  alias AcqdatCore.Repo
-
-  @spec create(%{description: any, metadata: any, name: any, org_id: any, parameters: any}) ::
-          {:error, %{error: %{optional(atom) => [any]}}} | {:ok, %{id: any, name: any, uuid: any}}
-  def create(params) do
-    %{
-      name: name,
-      description: description,
-      metadata: metadata,
-      parameters: parameters,
-      org_id: org_id,
-      generated_by: generated_by
-    } = params
-
-    verify_sensor_type(
-      SensorTypeModel.create(%{
-        name: name,
-        description: description,
-        metadata: metadata,
-        parameters: parameters,
-        org_id: org_id,
-        generated_by: generated_by
-      })
-    )
-  end
-
-  defp verify_sensor_type({:ok, sensor_type}) do
-    sensor_type = Repo.preload(sensor_type, :org)
-
-    {:ok,
-     %{
-       id: sensor_type.id,
-       name: sensor_type.name,
-       description: sensor_type.description,
-       metadata: sensor_type.metadata,
-       parameters: sensor_type.parameters,
-       org_id: sensor_type.org_id,
-       slug: sensor_type.slug,
-       uuid: sensor_type.uuid,
-       org: sensor_type.org,
-       generated_by: sensor_type.generated_by
-     }}
-  end
-
-  defp verify_sensor_type({:error, sensor_type}) do
-    {:error, %{error: extract_changeset_error(sensor_type)}}
-  end
-end
