--- apps/acqdat_api/lib/acqdat_api_web/validators/asset.ex
@@ -1,30 +1,0 @@
-defmodule AcqdatApiWeb.Validators.Asset do
-  use Params
-
-  defparams(
-    verify_asset(%{
-      owner_id: :integer,
-      creator_id!: :integer,
-      asset_category_id: :integer,
-      org_id!: :integer,
-      image_url: :string,
-      mapped_parameters: {:array, :map},
-      description: :string,
-      metadata: :map,
-      rgt: :integer,
-      properties: {:array, :string},
-      lft: :integer,
-      parent_id: :integer,
-      name: :string,
-      project_id!: :integer,
-      asset_type_id!: :integer
-    })
-  )
-
-  defparams(
-    verify_index_params(%{
-      page_size: :integer,
-      page_number: :integer
-    })
-  )
-end
