+++ apps/acqdat_core/lib/acqdat_core/entity-management/model/asset.ex
@@ -109,6 +109,7 @@ defmodule AcqdatCore.Model.EntityManagement.Asset do
         mapped_parameters: mapped_parameters,
         owner_id: owner_id,
         properties: properties
+
       }) do
     # NOTE: function Ecto.Changeset.__as_nested_set_column_name__/1 is undefined or private
     try do
@@ -138,6 +139,7 @@ defmodule AcqdatCore.Model.EntityManagement.Asset do
           mapped_parameters: mapped_parameters,
           owner_id: owner_id,
           properties: properties
+
         })
         |> create(:root)
         |> AsNestedSet.execute(Repo)
@@ -170,6 +172,22 @@ defmodule AcqdatCore.Model.EntityManagement.Asset do
         {:error, error.changeset}
     end
   end
+
+
+  # defp asset_struct(%{name: name, org_id: org_id, slug: slug, project_id: project_id}) do
+  #   %Asset{
+  #     name: name,
+  #     org_id: org_id,
+  #     project_id: project_id,
+  #     inserted_at: DateTime.truncate(DateTime.utc_now(), :second),
+  #     updated_at: DateTime.truncate(DateTime.utc_now(), :second),
+  #     uuid: UUID.uuid1(:hex),
+  #     slug: Slugger.slugify(slug),
+  #     properties: []
+  #   }
+  #   |> Repo.preload(:org)
+  #   |> Repo.preload(:project)
+  # end
 
   def asset_descendants(asset) do
     asset_descen_tree =
