+++ apps/acqdat_core/lib/acqdat_core/entity-management/schema/sensor.ex
@@ -9,6 +9,7 @@ defmodule AcqdatCore.Schema.EntityManagement.Sensor do
 
   use AcqdatCore.Schema
   alias AcqdatCore.Schema.EntityManagement.{Gateway, Organisation, Project, SensorsData}
+
   alias AcqdatCore.Schema.EntityManagement.{SensorType}
 
   @typedoc """
