--- apps/acqdat_core/lib/acqdat_core/schema/sensor_type_enum.ex
@@ -1,3 +1,0 @@
-import EctoEnum
-
-defenum(GeneratedBy, user: 0, asset: 1)
