+++ apps/acqdat_core/test/support/factory/hierarchy.ex
@@ -57,8 +57,7 @@ defmodule AcqdatCore.Factory.Hierarchy do
           properties: [],
           parent_id: -1,
           lft: -1,
-          rgt: -1,
-          project: insert(:project)
+          rgt: -1
         }
       end
